requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
# UPDATE: Bumped to satisfy LangChain 1.0.8 (requires >=2.7.4)
pydantic==2.9.2
# UPDATE: Bumped to match newer Pydantic
pydantic-settings==2.6.1

# LangChain & LangGraph (Pinned to stable 1.0 releases)
langchain==1.0.8
langchain-core>=1.0.5
langchain-openai==1.0.3
langchain-anthropic==1.1.0
langgraph==1.0.3
langgraph-checkpoint>=1.0.0
langgraph-prebuilt>=1.0.0
langgraph-sdk>=0.2.9

# OpenTelemetry (Updated for LiveKit compatibility)
opentelemetry-api>=1.34.0
opentelemetry-sdk>=1.34.0
opentelemetry-exporter-otlp>=1.34.0
opentelemetry-instrumentation-fastapi>=0.52b0

# Prometheus (Updated for LiveKit compatibility)
prometheus-client>=0.22.0

# LiveKit (Pinned to avoid backtracking)
livekit==1.0.19
livekit-api==1.0.7
livekit-agents==1.3.3
livekit-plugins-openai==1.3.3
livekit-plugins-deepgram==1.3.3
livekit-plugins-silero==1.3.3
livekit-plugins-elevenlabs==1.3.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pymongo==4.6.1

# Redis
redis==5.0.1
hiredis==2.3.2

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-multipart==0.0.6
jsonschema>=4.20.0

# AWS SDK for Secrets Manager (LocalStack)
boto3==1.34.37
botocore==1.34.37

# JWT for authentication
pyjwt==2.8.0
python-jose[cryptography]==3.3.0  # For RS256 JWKS validation (Zitadel)
bcrypt==4.1.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
